+++ src/components/utils.js
@@ -1,3 +1,5 @@
+import { debug } from '../debug'
+
 // requestAnimationFrame() polyfill by Erik Möller, Paul Irish, and Tino Zijdel.
 // https://gist.github.com/paulirish/1579671
 // eslint-disable-next-line no-extra-semi
@@ -67,6 +69,14 @@
     })
   }
 })()
+
+export function safeGetVar(variable) {
+  if (variable === undefined) {
+    if (debug) throw 'Variable is undefined.'
+    return false
+  }
+  return variable
+}
 
 export function safeExecuteFunc(f, ...a) {
   if (f && typeof f === 'function')